+++ src/main/java/com/sivitsky/ddr/service/VendorService.java
@@ -10,8 +10,8 @@ public interface VendorService {
 
     List<Vendor> listVendor();
 
-    void removeVendor(Integer id);
+    void removeVendor(Long id);
 
-    Vendor getVendorById(Integer id);
+    Vendor getVendorById(Long id);
 
 }
